+++ core/src/ten_manager/designer_frontend/public/locales/en-US/common.json
@@ -24,7 +24,7 @@
       "preferences": "Preferences",
       "installAll": "Install All"
     },
-    "officialSite": "Official site",
+    "officialSite": "Official Site",
     "poweredBy": "Powered by",
     "poweredByTenFramework": "Powered by <0></0>",
     "github": "GitHub",
